datapath: move fake/types.MTU to pkg/mtu/fake.MTU
